requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
joblib==1.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.1.2
openai==1.53.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.13
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
